Some fixes for better supporting the cases where we want to backport to a branch that hasn't had any releases made from it yet.
For example, we went ahead and made a v1.1.x branch off of master
before any releases were made for that branch (maybe we *should*
have just made an alpha release at that time?)

This change can just compare master with the release branch and
find any commits in master that aren't in the release branch.
With this I did get one false positive, but that could have something
to do with how the commit message was updated for that one PR
when it was manually backported.  This script was never able to be
*perfect* (hence 'suggest' in the name).