Remove hamogu/astrospec
Astrospec was meant as an alternative implementation for specutils at a time when that package was langushing. There is still stuff in astrospec that I plan to port to specutils at some point, but it's only a few functions and I don't maintain astrospec as a real "package" any longer.